+++ react/components/ButtonGroup.tsx
@@ -6,14 +6,12 @@ import style from './style.css'
 const ButtonGroup = ({ blockClass }: { blockClass: string }) => {
   const container__buttons = generateBlockClass(style.container__buttons, blockClass)
   const button__checkout = generateBlockClass(style.button__checkout, blockClass)
-  const button__btn = generateBlockClass(style.button__btn, blockClass)
   const button__cart = generateBlockClass(style.button__cart, blockClass)
 
   return (
     <>
       <div className={container__buttons}>
-        <a className={button__checkout} href="/">Check out</a>
-        <button className={button__btn}>Continúa Comprando</button>
+        <a className={button__checkout} >Continúa Comprando</a>
         {/* Checkout no es parte de spa de vtex io*/}
         <a href="/checkout/#/cart" className={button__cart}>Ver el carrito</a>
       </div>
