+++ react/components/ProductGroup.tsx
@@ -1,21 +1,10 @@
-import React from 'react'
+/* import React from 'react'
 import { generateBlockClass } from '@vtex/css-handles'
+import { ProductGroupProps } from './productGrouptype'
 import style from './style.css'
-interface ProductGroupProps {
-  products: {
-    name: string,
-    price: number,
-    imageUrls: string
-    id: string,
-    quantity: number,
-  }[];
-}
+
 const ProductGroup = (productGroup: ProductGroupProps, { blockClass }: { blockClass: string }) => {
-  const container__item = generateBlockClass(style.container__item, blockClass)
-  const container__img = generateBlockClass(style.container__img, blockClass)
-  const image = generateBlockClass(style.image, blockClass)
-  const container__item_info = generateBlockClass(style.container__item_info, blockClass)
-  const container__item_info_quantity = generateBlockClass(style.container__item_info_quantity)
+
   return (
     <>
 
@@ -35,10 +24,9 @@ const ProductGroup = (productGroup: ProductGroupProps, { blockClass }: { blockCl
           )
         })
       }
-
     </>
   )
 
 }
 
-export default ProductGroup
+export default ProductGroup */
